backend/requirements.txt
Django==3.2.3
django-filter==23.1
djangorestframework==3.14.0
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==7.0
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
python-dotenv==1.0.1
pyshorteners==1.0.1
